% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{arfimachanges}
\alias{arfimachanges}
\title{Prints changes to the package since the last update.  Started in 1.4-0}
\usage{
arfimachanges()
}
\description{
Prints changes to the package since the last update.  Started in 1.4-0
}
