% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARToPacf.R
\name{PacfToAR}
\alias{PacfToAR}
\title{Converts AR/MA coefficients from the PACF space to operator space}
\usage{
PacfToAR(pi)
}
\arguments{
\item{pi}{The AR/MA coefficients in PACF box-space}
}
\value{
The AR/MA coefficients in operator space.
}
\description{
Converts AR/MA coefficients from PACF box-space to operator space; usually
for internal use
}
\references{
Barndorff-Nielsen O. E., Schou G. (1973). "On the
parametrization of autoregressive models by partial autocorrelations."
Journal of Multivariate Analysis, 3, 408-419

McLeod A. I. , Zhang Y (2006).  "Partial autocorrelation parameterization
for subset autore- gression." Journal of Time Series Analysis, 27(4),
599-612
}
\author{
A. I. McLeod
}
\keyword{ts}
