% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityfcns.R
\name{print.arfima}
\alias{print.arfima}
\title{Prints a Fitted Object}
\usage{
\method{print}{arfima}(x, digits = max(6, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{A fitted \code{arfima} object}

\item{digits}{The number of digits to print}

\item{\dots}{Optional arguments.  See \code{\link{print}}.}
}
\value{
The object is returned invisibly
}
\description{
Prints a fitted \code{arfima} object's relevant details
}
\references{
Veenstra, J.Q. Persistence and Antipersistence:  Theory and
Software (PhD Thesis)
}
\author{
JQ (Justin) Veenstra
}
\keyword{ts}
