\name{print.tacvf}
\alias{print.tacvf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Prints a tacvf object.
}
\description{
Prints the output of a call to \code{tacvf} on an \code{arfima} object
}
\usage{
\method{print}{tacvf}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
The \code{tacvf} object.
}
  \item{\dots}{
Optional arguments.  See \code{\link{print}}.
}
}

\value{
The object is returned invisibly
}

\author{
Justin Veenstra
}


\seealso{
\code{\link{tacvf}}, \code{\link{plot.tacvf}}
}

\keyword{ ts}
