% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityfcns.R
\name{logLik.arfima}
\alias{logLik.arfima}
\title{Extract Log-Likelihood Values}
\usage{
\method{logLik}{arfima}(object, ...)
}
\arguments{
\item{object}{A fitted \code{arfima} object}

\item{\dots}{Optional arguments not currently used.}
}
\value{
A vector of log-likelihoods, one for each mode, is returned, along
with the degrees of freedom.
}
\description{
Extracts log-likelihood values from a \code{arfima} fit.
}
\details{
Uses the function \code{DLLoglikelihood} from the package
\code{\link{ltsa}}.  The log-likelihoods returned are exact up to an
additive constant.
}
\references{
Veenstra, J.Q. Persistence and Antipersistence:  Theory and
Software (PhD Thesis)
}
\seealso{
\code{\link{AIC.arfima}}
}
\author{
JQ (Justin) Veenstra
}
\keyword{ts}
