\name{print.arfima}
\alias{print.arfima}

\title{
Prints a Fitted Object
}
\description{
Prints a fitted \code{arfima} object's relevant details
}
\usage{
\method{print}{arfima}(x, digits 
= max(6, getOption("digits") - 3), ...) #
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A fitted \code{arfima} object
}
  \item{digits}{
The number of digits to print
}
  \item{\dots}{
Optional arguments.  See \code{\link{print}}.
}
}

\value{
The object is returned invisibly
}
\references{
Veenstra, J. and McLeod, A. I. (Working Paper).
The arfima R package: Exact Methods for Hyperbolic Decay Time Series

}
\author{
Justin Veenstra
}

\keyword{ ts}
