% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tacf.R
\name{print.tacvf}
\alias{print.tacvf}
\title{Prints a tacvf object.}
\usage{
\method{print}{tacvf}(x, ...)
}
\arguments{
\item{x}{The \code{tacvf} object.}

\item{\dots}{Optional arguments.  See \code{\link{print}}.}
}
\value{
The object is returned invisibly
}
\description{
Prints the output of a call to \code{tacvf} on an \code{arfima} object
}
\seealso{
\code{\link{tacvf}}, \code{\link{plot.tacvf}}
}
\author{
JQ (Justin) Veenstra
}
\keyword{ts}
